+++ src/server.h
@@ -2572,7 +2572,7 @@ void listTypeReleaseIterator(listTypeIterator *li);
 void listTypeSetIteratorDirection(listTypeIterator *li, unsigned char direction);
 int listTypeNext(listTypeIterator *li, listTypeEntry *entry);
 robj *listTypeGet(listTypeEntry *entry);
-void listTypeGetValue(listTypeEntry *entry, unsigned char **vstr, unsigned int *vlen, long long *lval);
+void listTypeGetValue(listTypeEntry *entry, unsigned char **vstr, size_t *vlen, long long *lval);
 void listTypeInsert(listTypeEntry *entry, robj *value, int where);
 void listTypeReplace(listTypeEntry *entry, robj *value);
 int listTypeEqual(listTypeEntry *entry, robj *o);
